:recycle: Split utils.functions into json and loop submodules
Separate the Django-boundary JSON helpers (model_to_json, json_to_model) from
the pure loop helpers (loopfirst, looplast, loopfirstlast) into
functions/json.py and functions/loop.py. functions/__init__.py re-exports all
five, so `from django_boost.utils.functions import ...` is unchanged; no
behavior change. This lets the pure loop helpers be typed cleanly while the
JSON helpers (django-stubs-dependent) are typed later.

Refs: https://github.com/ChanTsune/django-boost/issues/164
